checking user not none

+++ sd/views.py
@@ -18,18 +18,16 @@ def explore(request):
         print_state(request)
         posts = Post.objects.filter(Q(visibility=1 ) & (Q(unlisted=0) | Q(unlisted=False)))
         results = paginated_result(posts, request, "feed", query="feed")
-        if authenticated(request):
-            return render(request, 'sd/main.html', {'current_user': get_current_user(request), 'authenticated': True, 'results': results})
-        else:
-            return render(request, 'sd/main.html', {'current_user': None, 'authenticated': False, 'results': results})
+        is_authenticated = authenticated(request)
+        user = get_current_user(request) if is_authenticated else None
+        return render(request, 'sd/main.html', {'current_user': user, 'authenticated': is_authenticated, 'results': results})
     else:
         return HttpResponse(status_code=405)
 
 def feed(request):
     if valid_method(request):
-        print_state(request)
-        if authenticated(request):
-            user = get_current_user(request)
+        user = get_current_user(request)
+        if authenticated(request) and user:
             own_posts = Post.objects.filter(Q(author_id=user.uuid))
             pub_posts = Post.objects.filter(Q(visibility=1) & Q(unlisted=0))
             all_posts = own_posts | pub_posts
@@ -44,8 +42,8 @@ def feed(request):
 def account(request):
     if valid_method(request):
         print_state(request)
-        if authenticated(request):
-            user = get_current_user(request)
+        user = get_current_user(request)
+        if authenticated(request) and user:
             page = 'sd/account.html'
             return render(request, page, {'current_user': user, 'authenticated': True})
         else:
@@ -57,8 +55,8 @@ def account(request):
 def search(request):
     if valid_method(request):
         print_state(request)
-        if authenticated(request):
-            user = get_current_user(request)
+        user = get_current_user(request)
+        if authenticated(request) and user:
 
             # Get all authors
             all_authors = Author.objects.exclude(username=user)
@@ -205,8 +203,9 @@ def register(request):
 def logout(request):
     if valid_method(request):
         print_state(request)
-        if authenticated(request):
-            print("CONSOLE: Logging out "+get_current_user(request).username)
+        user = get_current_user(request)
+        if authenticated(request) and user:
+            print("CONSOLE: Logging out "+ user.username)
             try:
                 request.session['authenticated'] = False
                 request.session.pop('auth-user')
@@ -227,10 +226,10 @@ def friendrequest(request):
         if request.method == "GET":
             return HttpResponse(status_code=405)
 
-        if not authenticated(request):
+        user = get_current_user(request)
+        if not authenticated(request) or not user:
             print("CONSOLE: Redirecting from friendrequest because no one is logged in.")
             return redirect('login')
-        user = get_current_user(request)
         data = json.loads(request.body)
         target = Author.objects.get(username=data['target_author'])
         relationship, obj = get_relationship(user, target)
@@ -327,11 +326,11 @@ def friendrequest(request):
 def new_post(request):
     if valid_method(request):
         print_state(request)
-        if not authenticated(request):
+        user = get_current_user(request)
+        if not authenticated(request) or not user:
             print("CONSOLE: Redirecting from new_post because no one is logged in.")
             return redirect('login')
 
-        user = get_current_user(request)
         if request.method == "GET":
             form = NewPostForm()
             return render(request, 'sd/new_post.html', {'form': form, 'current_user': user, 'authenticated': True})
@@ -385,7 +384,8 @@ def get_image(request, url):
 def edit_post(request, post_id):
     if valid_method(request):
         print_state(request)
-        if not authenticated(request):
+        user = get_current_user(request)
+        if not authenticated(request) or not user:
             print("CONSOLE: Redirecting from edit_post because no one is logged in.")
             return redirect('login')
 
@@ -417,9 +417,10 @@ def edit_post(request, post_id):
 @csrf_exempt
 def delete_post(request, post_id):
     if request.method == "DELETE":
-        if authenticated(request):
+        user = get_current_user(request)
+        if authenticated(request) and user:
             post = Post.objects.get(uuid=post_id)
-            if post.author.uuid == get_current_user(request).uuid:
+            if post.author.uuid == user.uuid:
                 post.delete()
                 print("CONSOLE: Post deleted successfully.")
             else:
